% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.CollapseDataset.R
\name{GSEA.CollapseDataset}
\alias{GSEA.CollapseDataset}
\title{Maps user supplied identifiers to Gene Symbols}
\usage{
GSEA.CollapseDataset(dataplatform, gct, collapse.mode)
}
\arguments{
\item{dataplatform}{Dataframe returned by `GSEA.ReadCHIPFile`}

\item{gct}{Dataframe returned by `GSEA.Gct2Frame` after processing to remove unnecessary GCT header fields}

\item{collapse.mode}{Method for collapsing the dataset, accepts "max", "median", "mean", "sum", (default: NOCOLLAPSE)}
}
\value{
Gene by sample matrix converted into the namespace specified by the applied CHIP file
}
\description{
`GSEA.CollapseDataset` applies a .CHIP file to the dataset to gene symbols using a specified collapse function
}
\details{
Internal `GSEA` function invoked if collapse.dataset == TRUE.
}
\keyword{internal}
