% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.ConsPlot.R
\name{GSEA.ConsPlot}
\alias{GSEA.ConsPlot}
\title{Plots a heatmap of a consensus matrix}
\usage{
GSEA.ConsPlot(V, col.names, main = " ", sub = " ", xlab = " ",
  ylab = " ")
}
\description{
`GSEA.ConsPlot` plots heatmaps of the consensus matrix from leading edge analysis
}
\details{
Invoked by `GSEA.Analyze.Sets` to plot the consense matrix of the leading edge analysis.
}
\keyword{internal}
