% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.decompose.gs.R
\name{GSEA.decompose.gs}
\alias{GSEA.decompose.gs}
\title{Constructs a gene weight network from HumanNet v2 for a gene #' set and uses this network to decompose the gene set into 
subnetworks}
\usage{
GSEA.decompose.gs(
  gsdb,
  gs.names,
  gs.desc,
  gene.labels,
  gs.size.threshold.min,
  msigdbversion
)
}
\description{
`GSEA.decompose.gs` extracts subnetworks from gene sets to perform more granular subgraph level enrichment analysis
}
\details{
Internal `GSEA` function.
}
\keyword{internal}
