% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.EnrichmentScore.R
\name{GSEA.EnrichmentScore}
\alias{GSEA.EnrichmentScore}
\title{Computes the enrichment score of a gene set}
\usage{
GSEA.EnrichmentScore(gene.list, gene.set, weighted.score.type = 1,
  correl.vector = NULL)
}
\description{
`GSEA.EnrichmentScore` computes the weighted GSEA score of gene.set in gene.list
}
\details{
Internal `GSEA` function.
Computes the weighted GSEA score of gene.set in gene.list.  The weighted score
type is the exponent of the correlation weight: 0 (unweighted =
Kolmogorov-Smirnov), 1 (weighted), and 2 (over-weighted). When the score type
is 1 or 2 it is necessary to input the correlation vector with the values in
the same order as in the gene list.  Inputs: gene.list: The ordered gene list
(e.g. integers indicating the original position in the input dataset) gene.set:
A gene set (e.g. integers indicating the location of those genes in the input
dataset) weighted.score.type: Type of score: weight: 0 (unweighted =
Kolmogorov-Smirnov), 1 (weighted), and 2 (over-weighted) correl.vector: A
vector with the coorelations (e.g. signal to noise scores) corresponding to the
genes in the gene list Outputs: ES: Enrichment score (real number between -1
and +1) arg.ES: Location in gene.list where the peak running enrichment occurs
(peak of the 'mountain') RES: Numerical vector containing the running
enrichment score for all locations in the gene list tag.indicator: Binary
vector indicating the location of the gene sets (1's) in the gene list
}
\keyword{internal}
