% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.Gct2Frame.R
\name{GSEA.Gct2Frame}
\alias{GSEA.Gct2Frame}
\title{Reads .GCT file into a data frame for processing}
\usage{
GSEA.Gct2Frame(filename = "NULL")
}
\description{
`GSEA.Gct2Frame` is a wrapper for `read.table` with specified parameters for reading in a .GCT file
}
\details{
Internal `GSEA` function.
}
\keyword{internal}
