% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.get.local.network.R
\name{GSEA.get.local.network}
\alias{GSEA.get.local.network}
\title{Constructs a gene weight network from HumanNet v2 for a gene set}
\usage{
GSEA.get.local.network(global.netowrk, set, score.type = "strength")
}
\description{
`GSEA.get.local.network` constructs a network weight table for the genes in a gene set
}
\details{
Internal `GSEA` function.
}
\keyword{internal}
