% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.get.network.R
\name{GSEA.get.network}
\alias{GSEA.get.network}
\title{Constructs a gene weight network from HumanNet v2}
\usage{
GSEA.get.network(
  msigdbversion,
  gene.labels,
  score.type = "strength",
  weighted.score.type,
  expression.matrix = NULL
)
}
\description{
`GSEA.get.network` constructs a network weight table for all genes
}
\details{
Internal `GSEA` function.
}
\keyword{internal}
