% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.HeatMapPlot.R
\name{GSEA.HeatMapPlot}
\alias{GSEA.HeatMapPlot}
\title{Produce Heat Map for Genes in Dataset}
\usage{
GSEA.HeatMapPlot(V, row.names = F, col.labels, col.classes,
  col.names = F, main = " ", xlab = " ", ylab = " ")
}
\description{
`GSEA.HeatMapPlot` plots a heatmap 'pinkogram' of a gene expression matrix including phenotype vector and gene, sample and phenotype labels
}
\details{
Internal `GSEA` function invoked if gsea.type="GSEA"
}
\keyword{internal}
