% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.HeatMapPlot2.R
\name{GSEA.HeatMapPlot2}
\alias{GSEA.HeatMapPlot2}
\title{Produces overlap heatmap for leading edge analysis}
\usage{
GSEA.HeatMapPlot2(V, row.names = "NA", col.names = "NA", main = " ",
  sub = " ", xlab = " ", ylab = " ", color.map = "default")
}
\description{
`GSEA.HeatMapPlot2` plots a heatmap of a gene set leading edge overlap matrix
}
\details{
Internal function invoked by `GSEA.Analyze.Sets` to plot heatmaps of leading edge overlaps.
}
\keyword{internal}
