% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.Network.EnrichmentScore.R
\name{GSEA.Network.EnrichmentScore}
\alias{GSEA.Network.EnrichmentScore}
\title{Computes the enrichment score of a gene set}
\usage{
GSEA.Network.EnrichmentScore(
  gene.list,
  gene.set,
  correl.vector = NULL,
  net.set,
  correl.weight
)
}
\description{
`GSEA.Network.EnrichmentScore` computes the weighted GSEA score of gene.set in gene.list using the local set specific network to calcualte the weights for genes in the gene set and the global netowrk to clculate weights for the genes not in the gene set
}
\details{
Internal `GSEA` function.
Computes the weighted GSEA score of gene.set in gene.list.  The weighted score
type is the exponent of the correlation weight, where the weight is prodived by the 
strength score of each gene in the local network for genes in the gene set or the 
strength score of genes in the global network for genes not in the gene set.  
Inputs: gene.list: The ordered gene list (e.g. integers indicating the original 
position in the input dataset) gene.set: A gene set (e.g. integers indicating the 
location of those genes in the input dataset). correl.vector: A vector with the 
coorelations (e.g. signal to noise scores) corresponding to the genes in the gene 
list. net.set: the netowrk scores for each gene in the gene set. correl.weight: the 
global network scores for all genes in the dataset. Outputs: ES: Enrichment score 
(real number between -1 and +1) arg.ES: Location in gene.list where the peak 
running enrichment occurs (peak of the 'mountain') RES: Numerical vector containing 
the running enrichment score for all locations in the gene list tag.indicator: 
Binary vector indicating the location of the gene sets (1's) in the gene list
}
\keyword{internal}
