% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.NormalizeCols.R
\name{GSEA.NormalizeCols}
\alias{GSEA.NormalizeCols}
\title{Standardize columns of a gene expression matrix}
\usage{
GSEA.NormalizeCols(V)
}
\description{
`GSEA.NormalizeCols` performs column-wise normalization of a gene expression matrix using the col.mean and standard deviation
}
\details{
Internal `GSEA` function.
}
\keyword{internal}
