% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.NormalizeRows.R
\name{GSEA.NormalizeRows}
\alias{GSEA.NormalizeRows}
\title{Standardize rows of a gene expression matrix}
\usage{
GSEA.NormalizeRows(V)
}
\description{
`GSEA.NormalizeRows` performs row-wise adjustment of a gene expression matrix using the col.mean and standard deviation
}
\details{
Internal `GSEA` function.
}
\keyword{internal}
