% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.ReadCHIPFile.R
\name{GSEA.ReadCHIPFile}
\alias{GSEA.ReadCHIPFile}
\title{Read gene symbol mappings from .CHIP file}
\usage{
GSEA.ReadCHIPFile(file = "NULL")
}
\description{
`GSEA.ReadCHIPFile` is a wrapper for `read.table` with specified parameters for reading in a .CHIP file
}
\details{
Internal `GSEA` function.
}
\keyword{internal}
