% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.ReadClsFile.R
\name{GSEA.ReadClsFile}
\alias{GSEA.ReadClsFile}
\title{Read phenotype assignments from .CLS file}
\usage{
GSEA.ReadClsFile(file = "NULL")
}
\description{
`GSEA.ReadClsFile` reads a class vector CLS file to define dataset relationships
}
\details{
Read in a class vector CLS file to defines phenotype and class label
vectors for the samples in a gene expression file (GCT format).
}
\keyword{internal}
