% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.Threshold.R
\name{GSEA.Threshold}
\alias{GSEA.Threshold}
\title{Set thresholds for expression matrix}
\usage{
GSEA.Threshold(V, thres, ceil)
}
\description{
`GSEA.Threshold` sets threshold and ceiling values to pre-process gene expression matrix
}
\details{
Internal `GSEA` function.
}
\keyword{internal}
