% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.VarFilter.R
\name{GSEA.VarFilter}
\alias{GSEA.VarFilter}
\title{Variation filter pre-processing for gene expression matrix}
\usage{
GSEA.VarFilter(V, fold, delta, gene.names = "NULL")
}
\description{
`GSEA.VarFilter` apply a variation filter to expression matrix
}
\details{
Unused function.
}
\keyword{internal}
