% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.write.gct.R
\name{GSEA.write.gct}
\alias{GSEA.write.gct}
\title{Write output .GCT files}
\usage{
GSEA.write.gct(gct, filename)
}
\description{
`GSEA.write.gct` produces properly formatted .GCT files of leading edge subsets
}
\details{
Internal `GSEA.Analyze.Sets` function.
}
\keyword{internal}
